Big refactor of the utilities, now includes a CIMS.utils module
