+++ CIMS/utils/model_description/query.py
@@ -1,0 +1,22 @@
+from CIMS.readers.reader_utils import find_first_index
+from ..model_description import column_list as COL
+from ..parameter.parse import is_year
+
+
+def get_node_cols(mdf, first_data_col_name=COL.branch):
+    """
+    Returns list of column names after `first_data_col_name` and a list of years that follow
+    """
+    node_col_idx = find_first_index(mdf.columns,
+                                    lambda cn: first_data_col_name.lower() in cn.lower())
+
+    relevant_columns = mdf.columns[node_col_idx:]
+
+    year_or_not = list(map(is_year, relevant_columns))
+    first_year_idx = find_first_index(year_or_not)
+    last_year_idx = find_first_index(year_or_not[first_year_idx:],
+                                     lambda b: not b) + first_year_idx
+    # list(...)[a:][:b] extracts b elements starting at a
+    year_cols = mdf.columns[node_col_idx:][first_year_idx:last_year_idx]
+    node_cols = mdf.columns[node_col_idx:][:first_year_idx]
+    return node_cols, year_cols
